Add a mechanism to check whether a user is watching a video or not
Now it's possible to know if a user is currently watching a video by checking
the boolean attribute 'currentlyWatching' from Profile.

'currentlyWatching' is True at the moment in which the player view is called
and False when the user leaves the page.

Known bug: This implementation is not perfect. In case of a browser crash or
computer shutdown, 'currentlyWatching' is going to stay True. To avoid this
inconsistent state, the app could reset this attribute to false after a certain
amount of time. Another aproach could be to periodically send AJAX requests
instead of sending one when the user leaves. As we're developing a prototype, this
detail is going to be fixed later in order to continue with the implementation
of main features.